backend/requirements.txt
fastapi
asyncpg
uvicorn
gunicorn
python-dotenv
pydantic